+++ packages/fictoan-react/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix `Accordion` chevron position bug
 - Remove redundant `label` prop from `Badge`
 - Add `zIndex` as a prop for `Drawer` for easy, dynamic positioning
+- Tweak `Badge` delete button sizes
 
 ## 1.11.14
 - Refactor `Drawer` to behave more like the `Modal` for open and close behaviour ⚠️ BREAKING CHANGE